[qca-nss-ecm] Limit ECM database connection count to conntrack max
By limiting this connection count, we will prevent ECM to cause
out of memory in the system. ECM has a capability to accelerate
connections which are not in the conntrack database for some
special test scenarios, like bridge PPPoE. To allow this capability
for all the connections cause too much memory usage from the system's
memory. So, it is better to limit the total number of connections to the
conntracks max value.

Change-Id: I5f1c3e458036ee5490ac760043ae942e21bd7e84
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>