[qca-nss-ecm] Create DB connection only for supported non-ported flows
There is no need to create database connections for the
unsupported non-ported (other than UDP or TCP) flows. We already
do not accelarate them. Creating database connection for them
consumes memory and also causes to send unnecessary ARP and Neighbor
discovery messages from the ECM if the L2 address of the hosts are not
already resolved in the neigbor table. For example, there is no need
to create connections for the ping packets.

When we send the neighbor solicitation message to resolve the L2
address, we need to send these packets with the link local address
of the interface as the source address.

Change-Id: Ia53572a4098d3b8f5a2913a4cc78d991f44d0e60
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>