[qca-nss-ecm] Update conntrack timeout based on host's timer
Instead of updating the conntrack timer based on the
sync tick difference of NSS fw, update it based on the
ECM defunct timer of that connection. In each stats sync,
the conntrack entry's timeout is updated with a time value
which is the elapsed time since the last sync of that
connection.

Change-Id: I1e4e0fb77c60ac945f96588298d91db4a974f983
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>