[qca-nss-ecm] Conntrack cleans up DB when route event happens
nf_ct_iterate_cleanup function can destroy the conntrack
entries, if we return non-zero value in this iteration
callback implemented in ECM. Instead of destroying the
ECM database entries in this callback in ECM, it is better
to let the netfilter conntrack subsystem destroys the conntrack
entries. ECM will already receive these conntrack destroy
events and destroy the entries which are in our database.

Change-Id: Ibb2acf18ed4d06a51dbdbab4c2f533f5cc80eec6
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>