[qca-nss-ecm] Defunct the connections when PCC is loaded/unloaded
Regeneration of a connection causes issues,
when the regeneration process is handled with a packet
coming from the opposite direction of the connection.
This results to send arp requests to an IP address
through wrong interface while re-creating the interface
hierarchy. So, instead of regenerating the connections,
it is safer to defunct all the connections.

Change-Id: Ice6b73d6b0f3149b5eaa8bc816311b22cfe3d610
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>