[qca-nss-ecm] Fix SFE conntrack timeout update issue
Do not update the conntrack timeout when receive the final
sync message as a response of an ECM initiated destroy request,
if there is not any activity for that connection in the SFE driver.
(e.g: no tx or rx traffic)

This is not done for NSS frontends, because for the NSS the
elapsed timeout is calculated with a seperate function based on the
ECM database timer. In that function there is a timer group check, and
in this situation since the connection is not in a timer group,
that function already tells the sync callback to not update the conntrack.

Change-Id: I51b04d065fc4d73c6a2461130c974fcdbb38cfbc
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>