[qca-nss-ecm] Add protocol entry to feci
We didn't have protocol in feci but we had it in ci. While establishing
interface ci has not been initialized with the protocol number. We need
the protocol number to identify inner/outer dynamic interface type for
GRE like flows. To achieve this protocol number field is added to feci
and initialized before interface establish.

Change-Id: Ie5a883fac5dc864d28b5e4f8ae3068b3ba1dc8b2
Signed-off-by: Swaraj Sha <ssha@codeaurora.org>