[qca-nss-ecm] Fixed ref count issue caused by database timer callback
ecm_db_connection_defunct_callback is called when a connection's
timer expires. The callback first defuncts the connection and
assumes that the defunct is success. But the defunct can fail
and ECM retries to defunct, which sets the state of the connection
to MODE_ACCEL, or this defunct can happen while waiting an
acceleration response from NSS in which the state is
MODE_ACCEL_PENDING. So, the defunct callback shouldn't release the
last connection instance in these states.

Change-Id: I7a41a2c5c01be47e8d31b95c61ddab31440726ff
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>