[qca-nss-ecm] Use the correct destroy mechanism in decel pending state for multicast
There is a situation to call the destroy function for a connection
before that connection is fully accelerated. While it is in accel
pending state which means waiting for the response from the acceleration
engine. In this situation, the destroy function just returns and waits
for the accel response. When the response comes, it should use the
same destroy mechanism it tried before. Otherwise the state of the
connection becomes invalid.

Change-Id: Ie1328fe0fb8269a08505f449a316ce258442103e
Signed-off-by: Suman Ghosh <sumaghos@codeaurora.org>