[qca-nss-ecm] Remove assert in OVS classifier
OVS classifier depends on netdev priv flags to identify
if the device is OVS bridge port.  When the interface which is
active in forwarding traffic is removed from OVS bridge
sync function in classifier may not be able to find OVS bridge port
in from and to to hierarchy list in CI.  Because of this
assert on from and to is removed in sync function.

Change-Id: I4549afe2f9ee77b766213d008a66efbfabfbdaed
Signed-off-by: Bhaskar Valaboju <bhaskarv@codeaurora.org>