[qca-nss-ecm] Accelerate tunnel interface traffic in OVS
When a tunnel interface is added to OVS bridge, routed
traffic (post tunnel packets) is decelerated by OVS
classifier.  Tunnel packets are routed from tunnel interface
and not from OVS bridge interface.  This case is not handled
properly in OVS classifier and it is denying acceleration.
This patch is identifying this case and setting OVS classifier
as not relevant.

Change-Id: I33e3cec85ebce308b1dbdd0a78318dd2a861fbe6
Signed-off-by: Bhaskar Valaboju <bhaskarv@codeaurora.org>