[qca-nss-ecm] Fill DSCP extension entry during TCP handshake
After moving the TCP conntrack confirmation and establish check
from default classifier to the frontend file, DSCP extension
entry couldn't be filled in the DSCP classifier, because
it was doing it during the handshake. Now, we don't allow
flows to go through the classifiers before the handshake is done.
Thus, we need to fill the DSCP extension entry in the frontend file
and set the accel rule QoS values in the DSCP classifier.

Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>
Change-Id: I5d0d0efa6181462839b9c9ae9a00757dd5d43831