[qca-nss-ecm] Fix MTU change handler for regeneration
In ECM's database, the interfaces are stored with the acceleration
engine interface number as well. This means the same netdevice
can have multiple interfaces instances in the database, if
we create inner/outer interfaces. So, we need to go through
all the interfaces and compare the device identification
number to find the interface for updating the MTU values
of the connections created on that interface.

Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>
Change-Id: Ie0478e70e28384015d593d1899b98f50fd648872