[qca-nss-ecm] Establish dual bridge with same MAC address
When there are 2 bridges in the system with the same MAC address,
only one of them is stored in the database. Because we use only
MAC address to hash the interface and look it up with the MAC address only.
We need to add the netdevice interface index to the lookup function,
so that both bridges can be added and looked up by comparing both
MAC address and interface identifier.

Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>
Change-Id: Ie710c6d2f5154e1cebd6003bd70bf5dbcf13dd1b