[qca-nss-ecm] Set DSCP remark fields during TCP handshake.
This is for an alternate setting of these values in the
DSCP classifier response. Setting the values after the
handshake, in the classifier, by reading them from each packet
impacts the performance.

The existing setting is still kept, because when the conntrack
is flushed, the TCP connection still remains established.
When ECM re-creates the connection, since the TCP
handshake is bypassed, the DSCP and QoS values of the rules are
set in the DSCP classifier by reading them from the packeti header.

Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>
Change-Id: I4c70795b7a7533733cd7b50a2fafd264958a34e2