[qca-nss-ecm] Update ecm classifier stats earlier
When the connection is defuncted by ECM, the defunct
timer group of that connection is reset. For the reset
timer groups we have a check in the frontend sync function
and it returns from the function without doing
further updates. Emesh classifier needs this final
sync message to update its other modules.

Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>
Change-Id: I96dfeb2169648a6d4457671aeeba69a11726b8d3