[qca-nss-ecm] Fix ipv4 node address look-up process
When the node address look-up fails, we send ARP request
with the node's IP address. If there is a gateway between the
two devices, we need to find that gateway and send the ARP
request with the gateway's IP address.
For the bridge ports, we need to use the master device,
while sending the ARP requests.

Change-Id: I2cf9953e884d9db637bb306bb7b64023089483e5
Signed-off-by: Murat Sezgin <quic_msezgin@quicinc.com>