[qca-nss-ecm] Check the timer group before defunct
There is a race between the conntrack timeout and the ECM's
connection defunct timout. Conntrack timeout calls the
make_defunct function. In this function, if the timer entry
was removed by the ECM's database defunct timer, we shouldn't
attempt to defucnt the connection in this function. It will be
handled by the defunct timer callback.
Moreover, there may still be a timer removal after we check the
timer entry. In that case, we should release the last reference,
if we succeeded to defunct the connection.

Change-Id: I2fc2562590b726a1441de8089544dde87a35e9e3
Signed-off-by: Murat Sezgin <quic_msezgin@quicinc.com>