[qca-nss-ecm] Fix IPv4 node address look-up function
If the interface which the node is connected to is a
bridge port, we should use the master device to look up
the MAC address in the neighbor table as well as sending
the ARP request. Because the neighbor table has the L3 interfaces
only which have the IP addresses.

When a masq iptables rule is applied to the bridge interface
as below, the bridge flows become routed flow and in the
node establish function we do MAC lookup instead of reading the
MAC addresses from the packet.

iptables -t nat -A postrouting_lan_rule -s
192.168.1.0/24 -d 192.168.1.0/24 -o br-lan -j MASQUERADE

Change-Id: I021abb82c90e2283d1636e01ec14d41486345ef7
Signed-off-by: Murat Sezgin <quic_msezgin@quicinc.com>