[qca-nss-ecm] Fix race between ECM defunct functions
When a connection is defuncted, in the response of the
destroy message we set the accel_mode of the connection to
defunct fail mode for a very short time until the connection
is completely removed. This mode should be different than the
other defunct fail mode to avoid the race between two defunct calls
(ECM database timeout and manually called make_defunct function).
In this accel mode, the timeout callback shouldn't remove the
last reference.

Change-Id: I40b20d5877af0dcd9e0c160deca1880c62b20860
Signed-off-by: Murat Sezgin <quic_msezgin@quicinc.com>