[qca-nss-ecm] Add external module call rate counter to PCC
PCC calls the external module every 1 second. The first packet
is always skipped, because between the PCC instance creation
and the packet process in this function is always less than 1 second.
So, we should always call the external module for the first packet.
A new rate_exceeds counter is also added for statistics of this classifier.

Change-Id: Iccc1094637c70ab34ff9b64c54e0c49057c28e49
Signed-off-by: Murat Sezgin <quic_msezgin@quicinc.com>