[qca-nss-ecm] Stop traversing when node is found
In the MAC based defunct function, stop traversing more
when the MAC is found in the hash table. In a heavy load
traffic, a node with this MAC address may be created again
and added back to the node table, but the connections on that
node may not be fully created yet.

Change-Id: I1c2e4a4c23c2f09823939a89907657cbe20f4230
Signed-off-by: Murat Sezgin <quic_msezgin@quicinc.com>