[qca-nss-ecm] Add VxLAN bridge refresh functionality.
VxLAN clients will be added to the bridge fdb table.
These entries should be refreshed only when the traffic is offloaded to the PPE.
This is done to keep the bridge fdb table of linux in synchronization with PPE.
When the traffic is offloaded to NSS, the bridge fdb table for vxlan is refreshed
in vxlanmgr with the help of NSS firmware.
When the traffic is offloaded to SFE, the bridge fdb table for vxlan is refreshed by host itself.

Change-Id: I33527c14ca92c41fc4ca57b2babd7aa133432ed7
Signed-off-by: Sneha Maganahalli <quic_smaganah@quicinc.com>