[qca-nss-ecm] Fix the conntrack wasn't created when firewall stopped
when firewall stopped and no other netns ref was in use,
the conntrack could be removed and no conntrack was created for the new flow.
but the ecm still create the rule for those flow.
once firewall start again, since no destory event for the existing rule,
these rules can't be destroyed any more.

to keep the basic conntrack alive, we need to add refs to the netns.

Change-Id: I4126b59ff55d8508661809d317a9b84885933111
Signed-off-by: Ken Zhu <quic_guigenz@quicinc.com>