[qca-nss-ecm] Fix vxlan netdev down handler.
Vxlan net device has two interface instances associated with it.
One to handler inner flows, and the second one to handle outer flows.

As a result of this, when hadling net device DOWN event for vxlan
netdevice, we need to free the connections on both the interface
instances instead of just clearing the connections on the very first
interface intsance that matches the vxlan netdevice's interface index.

Also added an API to check if a net device has multiple interface
instances associated with it.

Change-Id: If5be431b7c9c07ab7a896a865549b11f37e13995
Signed-off-by: Aniruddha Bhat Anemajalu <quic_aanemaja@quicinc.com>