[qca-nss-ecm] Remove TCP establish check while setting dscp fields
In some scenarios, it is possible that the dscp remarked packet
can come to ECM with the 3rd packet of the 3-way handshake. At that time,
since the connection is in established state, we skip this setting and
dscp remarking fields stay unconfigured.
One of the cases is that using connmark match for one direction traffic
and setting the DSCP field of this direction based on that match.
In this case, DSCP target will be called for the 3rd packet of the TCP handshake.

Change-Id: I84c090d3769a171a27a6aa83701990d622e4dd44
Signed-off-by: Murat Sezgin <quic_msezgin@quicinc.com>