[qca-nss-ecm]: select SFE as default ae on sawf rule match.
Select SFE as default ae type on SAWF rule match.
But if user selects a particular ae ex: ppe, ppe-vp or
ppe-ds then that ae is used for acceleration. If no ae is
selected for flow or return direction rule then  SFE is
selected as default ae.

In the current code design to switch ae if flow and return
direction flows match with different ae then we do not do
ae switch.
Hence in scenarios where sawf rule matches for
flow direction and sal-rule output selects ppe-vp as selected ae,
but due to no rule match for return direction default ae is
considered as the selected ae by emesh-sp.
In this case we cannot consider return direction ae as SFE,
because then flow and return direction will select
different ae which is not supported by ae
switch functionality.

Change-Id: Ifd6bd21c74a48a0d3fec5d633f5a513eaca7b638