[qca-nss-ecm] Fix connection remove issue in flush and defunct case
This is a very rare case. Here is the path which ends up this state:

1) ECM sends accel rule to SFE (accel_mode = ACCEL_PENDING).
2) SFE flushes the rule before ECM receives the accel ACK and
   flush_happened is set to true.
3) Defunct is called for the same connection and is_defunct is set to true.
4) ECM receives the ACK of the acceleration rule. Since flush_happened was set,
   acceleration callback function set the accel_mode to DECEL.
5) There is no way anymore to defunct this connection when the
   accel_mode is DECEL and is_defunct is true.

To avoid this issue, accel_mode should be set to one of the FAIL modes,
so that the defunct callback can release the last reference when connection timeouts.

Change-Id: I989297ae4ccddf49f0bbbfc3f7c3cedf76e9d374
Signed-off-by: Murat Sezgin <quic_msezgin@quicinc.com>
(cherry picked from commit cbd90a90a8ed42b267e5b0427757468bf5a63886)