[qca-nss-ecm]: TCP traffic not getting priotized for SAWF.
The fix include the revert of the 3 link mlo changes for sawf
classifier.
Earlier, we were just enabling flow handle for sender type
and return handle for receiver type. In case of TCP traffic,
it was over writing the skb->mark with invalid fields.
With the changes, we can see the TCP priortization.

Change-Id: Ifff776c725b3ed7a0fbd6cb337a684d79c9ee216