[qca-nss-ecm]: Fix MSCS classification failing for TCP flows.
We are seeing following 2 issues with the current
code:

CASE 1:
Currently, we were waiting for 4 slow packets assuming that
atleast one slow ul packet will arrive with TID value.
But, in the test case, we saw that the tid arrives latest by
20 slow ul packets. Hence following are the changes:

 - If scs is enabled, we will continue to wait for 4 slow packets.
 - Else, we will deny acceleration for 20 slow uplink packets.

CASE 2:
We were seeing that, when the front-end switch happens from ppe to sfe,
few slow packets in between the process makes the classifier state
as no relevance. Hence, after the front end accel mode is in valid state,
still, we see that the classifier reattempt was not happening.
To encounter the issue,We will not make classifier state as no relevance
if ppe updates the front-end state as invalid. Instead, will wait for the
packet arriving with the relevant TID value.

Change-Id: Ie0d9a473683bea6978ad28f58790d5f66fd8b78a
Signed-off-by: Swati Singh <quic_swasing@quicinc.com>