[qca-nss-ecm] Fix to limit the ECM DB entries to the desired/configured value for SFE.
When the no. of connections to be SFE offloaded are set using the
appropriate
params in the ECM front end file, qca-nss-ecm, could observe that some
extra DB entries are also seen as part of ECM DUMP whose accel state is
determined to be "-4". Reason behind the same is that the pending flows
are not accounted properly and thus the flows when ran simulatenously
is leading to this problem.

Following commit will make sure that the ECM DB entries are limited
to the configured value as set in the ECM front file,
/etc/init.d/qca-nss-ecm by ensuring that the pending accel flows are
properly accounted during feci creation for the flows.

Signed-off-by: Vishnu Vardhan Bantanahal <quic_vishvard@quicinc.com>
Change-Id: I6e05d7e5aa9d7a7b61070605a80cbb0778f8acb9