[qca-nss-ecm] STC Mark Set Logic Fixes
This patch resolves two issues with the logic in the stc_mark_set() function:
1. An if check for if the classifier type was !invalid was instigating function bailout.
Since this function is called twice per connection, this was resulting in the second call
always bailing out, since the previous call set one of these classifier type parameters.

To resolve the issue this check was changed to !invalid and !ifli. This means that the code will
not bail if any service classes were set by STC.

2. The cemi parameters set in the sawf_mark_set() function were causing the sub to be called everytime before the add.
This is because the parameters were being changed prior to being checked against their previous values.

This function cannot be moved to after the sync message, since logically sync should not be called
until after the connection instances have been updated. Local variables are set to solve this issue

Change-Id: Ia5739541c137c59c92da07b05d20076cd7f0e8c1
Signed-off-by: John Sanli <quic_jsanli@quicinc.com>