[qca-nss-ecm] Fix race condition in FSE add at connection accel.
There is a potential race condition here, because of feci lock release before adding FSE
rule. After making connection into MODE_ACCEL, till the time FSE is added, destroy might come
(via defunct or accel ceased cases) which make one stale entry in FSE.

To avoid this, check if connection is in decel mode or not. If it is MODE_DECEL, delete FSE
entry. This is a WAR as FSE SFE path does not have any robust error handling and have to
redesign it in future.

Change-Id: Ida57e37e90a54afbbf089e4eab3f1cdcacf70827
Signed-off-by: Parikshit Gune <quic_pgune@quicinc.com>