[qca-nss-ecm] Fix STC MSDUQ Accounting in DSCP/VLAN + MLO Case
When a flow is accelerated ecm's sawf_process function will assign metadata
to particular flows if a dscp remarking or vlan pcp value is present.

STC's mark set function queries MSDUQ in two directions, as the STC rule is applied
to one direction at a time. When the query_msduq function was called
for the direction with an invalid service class, default msduq was being returned.
However, in the DSCP/VLAN cases the metadata associated with the flow by per the process function
is returned. This patch circumvents this assumption by simply assuming the default msduq
should be used for all invalid service class queries.

Change-Id: Idc71f8f9561405a6a19ba2c17a6b5021ecba7a13
Signed-off-by: John Sanli <quic_jsanli@quicinc.com>