[qca-nss-ecm] Compare protocol value during connection match
- Compare protocol value for connection match during
defunct. This is to ensure we do not end up flushing
rule which do not match the protocol value.

- Existing logic of matching the protocol with
mask (proto & mask) == mask) leads to connection match
even when the protocol from the defunct event and the
connection instance are different.
- For example, proto=47, proto_mask=1 would result into
flushing the ECM rule with protocol 47 as (0x2F & 0x1 = 1)
- Similarly for proto=47, proto_mask=6, will result in
match (0x2F & 0x6 = 0x6)
- There could be other combinations that may also fail.

Change-Id: I4d58bda5aab356e37a1723110eb56078064b7463
Signed-off-by: Pritam Jha <quic_pritamku@quicinc.com>