[qca-nss-ecm] Overwrite top iface to ppe_vxlan only for outer flow
Overwrite the top interface id in ECM rule for PPE with vxlan
dev id only if top and bottom iface are same in one direction,
case with VxLAN outer flow. This is required as linux vxlan
netdevice details are not available with PPE.
If the inner flow is routed before coming to vxlan, top interface
could be other than vxlan, like br-lan. If we overwrite the top
interface as vxlan for inner flow also then ppe will not be able
to determine the nexthop.

Change-Id: I16b32c87dc9d3903d3e8da48dbef3650f608093f
Signed-off-by: Pritam Jha <pritamku@qti.qualcomm.com>