[qca-nss-ecm] Move the conntrack confirm check after connection creation
This is done only for UDP connections. With iperf3 tool, when uni-direction
flow is created, the first packet is skipped in ECM and the connection is
created with the single packet coming from the server side. This causes
the ECM connection creation in the reverse direction of the conntrack entry.
This is not an issue for ECM and AEs since they are bi-directional
AEs and managers, but it causes some confusions for the test teams. So,
skipping the flow after the connection is created in ECM will guarantee
creating the connection in the original direction.

Change-Id: I810f9a10bdba98a84e7fc7ecc38d961c7860dc88
Signed-off-by: Murat Sezgin <msezgin@qti.qualcomm.com>