[qca-nss-ecm] Dont send neighbor solicitation frame on Tunnel netdevices as source
During network restart with traffic running in background if the tunnel
is created using ip command the interface is not reconfigured but the
underlying WAN interface might get reconfigured. There would be some
delay in assignement of ipv6 address on the WAN port. During which we
cant  determine if its a local IP address leading to Tx of NS frame on
tunnel interface.

Adding checks to send NS frames on tunnel interface if the protocol is
not the same type as that of tunnel.

Change-Id: Ic3ad543ec51850d5c57aca152c4feae532c10276
Signed-off-by: Karthik T S <quic_kartikts@quicinc.com>
(cherry picked from commit 2e68aa8662588c125875e49aeea657301c17434d)