[qca-nss-ecm] Don't stop traversing the node chain in FDB event handler
Each node instance is bound to an interface. If there are both
bridge and routed flows from a client which is connected to a
bridge port, there are two node instances created for the MAC address
of this client. One is bound to bridge port for bridge connections,
the other one is bound to bridge interface for routed connections.

Change-Id: I0330ac6d07dbc17ca67510dc0e677f54347adb49
Signed-off-by: Murat Sezgin <msezgin@qti.qualcomm.com>
(cherry picked from commit f1de20b09a773806ade7ef359a7546cebd6f05df)