[qca-nss-ecm] Set accel mode to FAIL_DEFUNCT when is_defunct is true
- Update accel mode of connection instance to FAIL_DEFUNCT
if is_defunct is true while connection is getting accelerated
and fails for some reason in sfe path.
- Very few times it happens that while accel mode is
pending and defunct is called, both the flags is_defunct,
decelerate_pending are set to true for the connection.
- When this connection fails to accelerate, sfe rule
push failure in this case, the decelerate_pending is
cleared but is_defunct state is not taken care and
it remains true.
- While clearing the decel pending flag also check for
is_defunct and change the accel mode to failure so that
the connection does not get accelerated again and
removed eventually.

Change-Id: Ic8eb3db978eabbc6bfb2adca0b8dc182cc7bb14e
Signed-off-by: Pritam Jha <pritamku@qti.qualcomm.com>