[qca-nss-ecm] Fix connection regeneration logic
We need to regenerate the connections when we see the
packets in the same direction with the ECM connection instance.
Otherwise, ecm_front_end_ipv4_interface_construct_set_and_hold()
function cannot set the input interfaces for the interface hierarchy
re-construction correctly. We use "sender" to determine if the packet
is coming in the same direction with the connection instance. Then,
we normalize it again relative to the conntrack direction to use it
for setting the input parameters of the interface hierarchy
creation function.

Reproduction of the issue:
1. WANPC (192.168.2.1) -（WAN: 192.168.2.2)DUT(LAN: 192.168.1.1) - LANPC (192.168.1.10)
2. Use traffic generator to run a TCP flow from LAN PC to WAN PC
   Conntrack entry is created from 192.168.1.10 -> 192.168.2.1
3. repeat to defunct ECM ci (echo 1 > /sys/kernel/debug/ecm/ecm_db/defunct_all)
   until ECM creates flow in the conntarck's reply direction 192.168.2.1 to 192.168.1.10
4. ifconfig br-lan mtu 1506 => trigger regeneration
5. Connection will never be re-accelerated and stay in accel_mode=0

Change-Id: I66f34965d35225a8e11db15b5195b045c407c272
Signed-off-by: Murat Sezgin <msezgin@qti.qualcomm.com>